+++ apps/dashboard/components/layout/site-builder/navigation-publish-dropdown.tsx
@@ -7,87 +7,28 @@ import {
   DropdownMenuGroup
 } from '@/ui/dropdown-menu';
 import { Button } from '@/components/ui/button';
-import { ChevronDownIcon, ArrowRightIcon } from '@radix-ui/react-icons';
+import {
+  ChevronDownIcon,
+  ArrowRightIcon,
+  Pencil1Icon
+} from '@radix-ui/react-icons';
 import { motion } from 'framer-motion';
 import type { websites as Website } from '@prisma/client';
 import Link from 'next/link';
-import { getEnv, getTimeStamp } from '@/utils/index';
-import { publishSite, updateWebsite } from '@/utils/websites';
-import { toast } from 'sonner';
-import ModalDeleteSite from '@/components/modals/site/modal-delete-site';
-import { Dialog, DialogTrigger } from '@/components/ui/dialog';
-import { DialogContent } from '@radix-ui/react-dialog';
+import { cn, getEnv } from '@/utils/index';
+import PublishDropdownItems from './navigation-publish-dropdown-items';
+import { useState } from 'react';
 
 export default function SiteBuilderPublishModal({
   website
 }: {
   website: Website;
 }) {
+  const [websiteUrlEditor, setWebsiteUrlEditor] = useState(false);
+
   const env = getEnv() === 'production' ? 'https' : 'http';
-
-  // either update the website or publish it depending on if its been published before
-  const handleWebsitePublish = async () => {
-    if (website.hasBeenPublished) {
-      updateWebsite(website.websiteId, {
-        ...website
-      });
-      toast.success('Website updated successfully');
-    } else {
-      publishSite(website.websiteId);
-      toast.success('Website published successfully');
-    }
-  };
-
-  const dropdownOptions = [
-    {
-      name: 'lastUpdated',
-      html: (
-        <div className="flex flex-col gap-y-1 text-xs p-2">
-          <span className="flex items-center gap-x-1">
-            Last updated
-            <p className="bg-black-50 p-1 rounded">
-              {getTimeStamp(website?.lastUpdated)}
-            </p>
-          </span>
-          <div className="">By 'Unknown'</div>
-        </div>
-      )
-    },
-    {
-      name: 'separator'
-    },
-    {
-      name: 'publish-unpublish',
-      html: (
-        <div className="flex flex-col w-full gap-2 items-center font-inter mt-2 [&>div]:w-full">
-          <Dialog modal={true}>
-            <DialogTrigger
-              className="w-full"
-              asChild
-            >
-              <Button
-                variant="destructive"
-                className="p-2 group-hover:!bg-white group-hover:text-black text-xs !border-none"
-                rounded="sm"
-              >
-                <span className="w-full">Delete</span>
-              </Button>
-            </DialogTrigger>
-
-            <ModalDeleteSite siteId={website?.websiteId} />
-          </Dialog>
-          <Button
-            variant="secondary"
-            className="p-2 group-hover:!bg-white group-hover:text-black text-xs !border-none"
-            rounded="sm"
-            onClick={async () => await handleWebsitePublish()}
-          >
-            <span>{website?.hasBeenPublished ? 'Update' : 'Publish'}</span>
-          </Button>
-        </div>
-      )
-    }
-  ];
+  // only allow the user to modify the first part of the website url
+  const websiteUrlEditable = website?.websiteUrl?.split('.')[0];
 
   return (
     <DropdownMenu>
@@ -118,20 +59,44 @@ export default function SiteBuilderPublishModal({
         >
           <DropdownMenuLabel className="text-base font-normal flex flex-col gap-y-1 font-inter">
             <span className="text-xs text-white/60">
-              {website?.websiteName}
+              {website?.websiteName}: {websiteUrlEditor}
             </span>
             {website?.websiteUrl ? (
-              <Link
-                className="text-sm text-white group flex items-center gap-x-1 hover:underline"
-                href={`${env}://${website?.websiteUrl}`}
-                prefetch
-                target="_blank"
-              >
-                {website?.websiteUrl || 'Not Published'}
-                <span>
-                  <ArrowRightIcon className="relative size-3 opacity-0 group-hover:opacity-100 -left-2 group-hover:left-0 duration-300 transition-all" />
-                </span>
-              </Link>
+              <div className="flex items-center justify-between w-full">
+                {websiteUrlEditor ? (
+                  <input
+                    type="text"
+                    className="w-full text-white/80 bg-transparent focus-visible:outline-none placeholder:text-[10px]"
+                    value={websiteUrlEditable}
+                    onChange={(e) => {
+                      console.log(e.target.value);
+                    }}
+                  />
+                ) : (
+                  <Link
+                    className="text-sm text-white group flex items-center gap-x-1 hover:underline"
+                    href={`${env}://${website?.websiteUrl}`}
+                    prefetch
+                    target="_blank"
+                  >
+                    {website?.websiteUrl || 'Not Published'}
+                    <ArrowRightIcon className="relative size-3 opacity-0 group-hover:opacity-100 -left-2 group-hover:left-0 duration-300 transition-all" />
+                  </Link>
+                )}
+                <button
+                  onClick={() => {
+                    setWebsiteUrlEditor(!websiteUrlEditor);
+                    console.log(websiteUrlEditor);
+                  }}
+                >
+                  <Pencil1Icon
+                    className="cursor-pointer transition-colors duration-300"
+                    style={{
+                      color: websiteUrlEditor ? '#6725f2' : '#ffffff'
+                    }}
+                  />
+                </button>
+              </div>
             ) : (
               <span className="text-sm text-white">Not Published</span>
             )}
@@ -139,24 +104,7 @@ export default function SiteBuilderPublishModal({
         </motion.div>
         <DropdownMenuSeparator className="!bg-black-50" />
         <DropdownMenuGroup>
-          {dropdownOptions.map((option, index) =>
-            option.name === 'separator' ? (
-              <DropdownMenuSeparator
-                key={index}
-                className="!bg-black-50"
-              />
-            ) : (
-              <motion.li
-                key={option.name}
-                initial={{ opacity: 0, x: 0 }}
-                animate={{ opacity: 1, x: 0 }}
-                transition={{ duration: 0.2 }}
-                className="list-none w-full"
-              >
-                {option.html}
-              </motion.li>
-            )
-          )}
+          <PublishDropdownItems website={website} />
         </DropdownMenuGroup>
       </DropdownMenuContent>
     </DropdownMenu>
